Implementing MVP pattern for product form dialog.

+++ app/src/main/java/com/mithril/mobilegoldenleaf/models/Product.kt
@@ -13,7 +13,7 @@ import java.math.BigDecimal
 @Entity(foreignKeys = [ForeignKey(entity = Category::class, parentColumns = ["id"], childColumns = ["categoryId"], onDelete = SET_NULL)])
 class Product : Serializable {
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0
     @ForeignKey(entity = Category::class, parentColumns = ["id"], childColumns = ["categoryId"])
     var categoryId: Int = 0
     var brand: String? = null
@@ -23,7 +23,7 @@ class Product : Serializable {
     var isAvailable: Boolean = false
 
     @Ignore
-    constructor(id: Int, category_id: Int, brand: String, description: String, code: String, unit_cost: BigDecimal, is_available: Boolean) {
+    constructor(id: Long, category_id: Int, brand: String, description: String, code: String, unit_cost: BigDecimal, is_available: Boolean) {
         this.id = id
         this.categoryId = category_id
         this.brand = brand
